멤버 주간 기록 api 작업

+++ app/src/main/java/com/ssafy/neegongnaegong/presentation/navigation/AppNavigation.kt
@@ -71,7 +71,7 @@ object AppNavigation {
             data class MakeVote(val studyGroupId: Int) : Studies
 
             @Serializable
-            data class Record(val groupId: Int, val memberId: Int) : Studies
+            data class Record(val groupId: Long, val memberId: Long) : Studies
         }
 
         @Serializable
